XXX Don't upstream buildRustPackage: apply patches when fetching cargo deps
This fixes an issue with applying a patch to an upstream `Cargo.lock`,
where the patch would be applied to the build itself, but not when
downloading dependencies (thus ignoring dependency changes in
`Cargo.lock`).

Some commands had to be moved to the `postPatch` phase, so that
patches to `Cargo.lock` are taken into account.